more refactoring of filters
