+++ app/src/components/Filters/LabelFieldFilters.state.tsx
@@ -1,0 +1,153 @@
+import { selector, selectorFamily } from "recoil";
+
+import { Range } from "./RangeSlider";
+import * as atoms from "../../recoil/atoms";
+import * as selectors from "../../recoil/selectors";
+
+const getRange = (
+  get: GetRecoilValue,
+  path: string,
+  range: Range | DefaultValue
+): Range => {
+  const bounds = get(selectors.numericFieldBounds(path));
+  return bounds; // todo;
+};
+
+export const confidenceRangeAtom = selectorFamily<Range, string>({
+  key: "confidenceRangeAtom",
+  get: ({ get }) => {},
+  set: (path) => ({ get, set }, value) => {},
+});
+
+export const labelFilters = selector({
+  key: "labelFilters",
+  get: ({ get }) => {
+    const frameLabels = get(atoms.activeLabels("frame"));
+    const labels = {
+      ...get(atoms.activeLabels("sample")),
+      ...Object.keys(frameLabels).reduce((acc, cur) => {
+        return {
+          ...acc,
+          ["frames." + cur]: frameLabels[cur],
+        };
+      }, {}),
+    };
+    const filters = {};
+    for (const label in labels) {
+      const range = get(filterLabelConfidenceRange(label));
+      const none = get(filterLabelIncludeNoConfidence(label));
+      const include = get(filterIncludeLabels(label));
+      filters[label] = (s) => {
+        const inRange =
+          range[0] - 0.005 <= s.confidence && s.confidence <= range[1] + 0.005;
+        const noConfidence = none && s.confidence === undefined;
+        const isIncluded = include.length === 0 || include.includes(s.label);
+        return (inRange || noConfidence) && isIncluded;
+      };
+    }
+    return filters;
+  },
+});
+
+export const modalLabelFilters = selector({
+  key: "modalLabelFilters",
+  get: ({ get }) => {
+    const frameLabels = get(atoms.modalActiveLabels("frame"));
+    const labels = {
+      ...get(atoms.modalActiveLabels("sample")),
+      ...Object.keys(frameLabels).reduce((acc, cur) => {
+        return {
+          ...acc,
+          ["frames." + cur]: frameLabels[cur],
+        };
+      }, {}),
+    };
+    const hiddenObjects = get(atoms.hiddenObjects);
+    const filters = {};
+    for (const label in labels) {
+      const range = get(atoms.modalFilterLabelConfidenceRange(label));
+      const none = get(atoms.modalFilterLabelIncludeNoConfidence(label));
+      const include = get(atoms.modalFilterIncludeLabels(label));
+      filters[label] = (s) => {
+        if (hiddenObjects[s.id]) {
+          return false;
+        }
+        const inRange =
+          range[0] - 0.005 <= s.confidence && s.confidence <= range[1] + 0.005;
+        const noConfidence = none && s.confidence === undefined;
+        const isIncluded = include.length === 0 || include.includes(s.label);
+        return labels[label] && (inRange || noConfidence) && isIncluded;
+      };
+    }
+    return filters;
+  },
+  set: ({ get, set }, _) => {
+    const paths = get(labelPaths);
+    const activeLabels = get(atoms.activeLabels("sample"));
+    set(atoms.modalActiveLabels("sample"), activeLabels);
+    const activeFrameLabels = get(atoms.activeLabels("frame"));
+    set(atoms.modalActiveLabels("frame"), activeFrameLabels);
+    for (const label of paths) {
+      set(
+        atoms.modalFilterLabelConfidenceRange(label),
+        get(filterLabelConfidenceRange(label))
+      );
+
+      set(
+        atoms.modalFilterLabelIncludeNoConfidence(label),
+        get(filterLabelIncludeNoConfidence(label))
+      );
+
+      set(
+        atoms.modalFilterIncludeLabels(label),
+        get(filterIncludeLabels(label))
+      );
+
+      set(atoms.modalColorByLabel, get(atoms.colorByLabel));
+    }
+  },
+});
+
+export const sampleModalFilter = selector({
+  key: "sampleModalFilter",
+  get: ({ get }) => {
+    const filters = get(modalLabelFilters);
+    const frameLabels = get(atoms.modalActiveLabels("frame"));
+    const activeLabels = {
+      ...get(atoms.modalActiveLabels("sample")),
+      ...Object.keys(frameLabels).reduce((acc, cur) => {
+        return {
+          ...acc,
+          ["frames." + cur]: frameLabels[cur],
+        };
+      }, {}),
+    };
+    return (sample) => {
+      return Object.entries(sample).reduce((acc, [key, value]) => {
+        if (key === "tags") {
+          acc[key] = value;
+        } else if (value && VALID_LIST_TYPES.includes(value._cls)) {
+          acc[key] =
+            filters[key] && value !== null
+              ? {
+                  ...value,
+                  [value._cls.toLowerCase()]: value[
+                    value._cls.toLowerCase()
+                  ].filter(filters[key]),
+                }
+              : value;
+        } else if (value !== null && filters[key] && filters[key](value)) {
+          acc[key] = value;
+        } else if (RESERVED_FIELDS.includes(key)) {
+          acc[key] = value;
+        } else if (
+          ["string", "number", "null"].includes(typeof value) &&
+          activeLabels[key]
+        ) {
+          acc[key] = value;
+        }
+        return acc;
+      }, {});
+    };
+  },
+});
